+++ src/Zicht/Bundle/SolrBundle/Manager/Doctrine/Subscriber.php
@@ -10,6 +10,7 @@ use Doctrine\Common\EventSubscriber;
 use Doctrine\ORM\Event\LifecycleEventArgs;
 use Doctrine\ORM\Events;
 use Symfony\Component\DependencyInjection\ContainerInterface;
+use Zicht\Bundle\SolrBundle\Solr\QueryBuilder\Interfaces\Extractable;
 
 /**
  * Class Subscriber
@@ -20,7 +21,7 @@ class Subscriber implements EventSubscriber
      * @var ContainerInterface
      */
     protected $container;
-    
+
     /**
      * Construct the subscriber
      *
@@ -39,7 +40,7 @@ class Subscriber implements EventSubscriber
      */
     public function postPersist(LifecycleEventArgs $event)
     {
-        $this->container->get('zicht_solr.manager')->update($event->getEntity());
+        $this->callUpdate($event);
     }
 
 
@@ -51,7 +52,7 @@ class Subscriber implements EventSubscriber
      */
     public function preUpdate(LifecycleEventArgs $event)
     {
-        $this->container->get('zicht_solr.manager')->update($event->getEntity());
+        $this->callUpdate($event);
     }
 
     /**
@@ -75,5 +76,22 @@ class Subscriber implements EventSubscriber
             Events::preUpdate,
             Events::preRemove
         );
+    }
+
+    /**
+     * Calls the proper method in the Solr Manager to update or extract the document
+     *
+     * @param LifecycleEventArgs $event
+     */
+    private function callUpdate(LifecycleEventArgs $event)
+    {
+        $entity = $event->getEntity();
+        if ($entity instanceof Extractable && is_resource($entity->getFileResource())) {
+            $this->container->get('zicht_solr.manager')->extract($entity);
+
+            return;
+        }
+
+        $this->container->get('zicht_solr.manager')->update($entity);
     }
 }