+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 3.3 #
+* Added Solr Entity Manager to better manage updates of entities and their related entities
+
+# 3.2 #
+* Added Extract Query, uses SOLR's capability to index Document Formats as PDF, DOC, etc.
+
 # 3.1 #
 * Added SolrDataCollector to show request send from debug toolbar